monroe-nrepl-server-start: change directory before starting server.
In monroe-nrepl-server-start, change the directory
to (monroe-get-directory) before starting the server. This is because
clj requires that the server be started in the directory where the
deps.edn is, not in a subdirectory thereof.

Also, start the server with start-process rather than
async-shell-command. This is recommended for non-interactive use, and
it doesn't pop up a buffer. As a consequence,
monroe-nrepl-server-cmd-args must be a list.